Updates post tech review

+++ src/SFA.DAS.Approvals/Application/Fjaa/Queries/GetAgencyResult.cs
@@ -2,6 +2,6 @@
 {
     public class GetAgencyResult
     {
-        public int LegalEntityId { get; set; }
+        public long LegalEntityId { get; set; }
     }
 }
